Refactor registration methods to return Long ID instead of Boolean
- Updated DronesConfigController, DronesModelController, DronesRouteLibraryController, DronesServicesController, DronesTaskController, DronesWorkflowController, SysDictItemController, SysLogController, SysMenuController, and SysOrgController to change the return type of create methods from R<Boolean> to R<Long>.
- Modified corresponding service interfaces and implementations (DronesConfigService, DronesModelService, DronesRouteLibraryService, DronesServicesService, DronesTaskService, DronesWorkflowService, SysDictService, SysDictItemService, SysLogService, SysMenuService, SysOrgService, and SysRoleService) to return Long from register methods instead of Boolean.
- Changed entity ID type from Integer to Long in SysDict, SysDictItem, SysMenu, SysOrg, and SysRole entities.
- Updated SysOrgRepository to use Long as the ID type.
- Adjusted tests for SysRoleController and SystemUserController to handle Long ID returns.

+++ src/main/java/org/huazhi/system/sysorg/entity/dto/SysOrgQueryDto.java
@@ -20,5 +20,5 @@ public class SysOrgQueryDto {
     @QueryParam("endTime")
     private String endTime;
 
-    private Integer notId;
+    private Long notId;
 }
